+++ scalaz-7.1/src/main/scala/japgolly/scalajs/react/ScalazReact.scala
@@ -61,51 +61,48 @@ object ScalazReact {
   final type ReactS[S, A] = ReactST[Id, S, A]
   final type ReactST[M[_], S, A] = StateT[M, StateAndCallbacks[S], A]
 
+  /**
+   * M prefix indicates M in args.
+   * T prefix indicates we want the result lifted into an M.
+   */
   object ReactS {
-    @inline def apply[S, A](f: S => (S, A)): ReactS[S, A] = applyT[Id, S, A](f)
+    @inline def apply    [S,A](f: S => (S, A))       : ReactS[S,A]    = applyM[Id, S, A](f)
+    @inline def callback [S,A](a: A, c: OpCallbackIO): ReactS[S,A]    = callbackM[Id, S, A](a, c)
+    @inline def get      [S]                         : ReactS[S,S]    = gets(identity[S])
+    @inline def mod      [S]  (f: S => S)            : ReactS[S,Unit] = modM[Id, S](f)
+    @inline def ret      [S,A](a: A)                 : ReactS[S,A]    = retM[Id, S, A](a)
+    @inline def set      [S]  (s: S)                 : ReactS[S,Unit] = mod((_: S) => s)
 
-    @inline def ret[S, A](a: A): ReactS[S, A] = retT[Id, S, A](a)
+    @inline def applyT   [M[_],S,A](f: S => (S, A))         (implicit M: Applicative[M]): ReactST[M,S,A]    = applyM(s ⇒ M point f(s))
+    @inline def callbackT[M[_],S,A](a: A, c: OpCallbackIO)  (implicit M: Applicative[M]): ReactST[M,S,A]    = callbackM(M point a, c)
+    @inline def getT     [M[_],S]                           (implicit M: Applicative[M]): ReactST[M,S,S]    = get.lift[M]
+    @inline def getsT    [M[_],S,A](f: S => A)              (implicit M: Applicative[M]): ReactST[M,S,A]    = getsM(s ⇒ M point f(s))
+    @inline def liftR    [M[_],S,A](f: S ⇒ ReactST[M, S, A])(implicit M: Monad[M])      : ReactST[M,S,A]    = getT[M,S] flatMap f
+    @inline def modT     [M[_],S]  (f: S => S)              (implicit M: Applicative[M]): ReactST[M,S,Unit] = modM(s ⇒ M point f(s))
+    @inline def retM     [M[_],S,A](ma: M[A])               (implicit M: Functor[M])    : ReactST[M,S,A]    = getsM[M,S,A](_ ⇒ ma)
+    @inline def setM     [M[_],S]  (ms: M[S])               (implicit M: Functor[M])    : ReactST[M,S,Unit] = modM((_: S) ⇒ ms)
+    @inline def setT     [M[_],S]  (s: S)                   (implicit M: Applicative[M]): ReactST[M,S,Unit] = setM(M point s)
 
-    @inline def get[S]: ReactS[S, S] = gets(identity[S])
-
-    @inline def gets[S, A](f: S => A): ReactS[S, A] =
-      State.gets[StateAndCallbacks[S], A](s => f(s.s))
-
-    @inline def set[S](s: S): ReactS[S, Unit] = mod((_: S) => s)
-
-    @inline def mod[S](f: S => S): ReactS[S, Unit] = modT[Id, S](f)
-
-    @inline def callback[S, A](a: A, c: OpCallbackIO): ReactS[S, A] = callbackM[Id, S, A](a, c)
-
-    def applyT[M[_], S, A](f: S => M[(S, A)])(implicit F: Functor[M]): ReactST[M, S, A] =
+    def applyM[M[_], S, A](f: S => M[(S, A)])(implicit F: Functor[M]): ReactST[M, S, A] =
       StateT[M, StateAndCallbacks[S], A](sc => F.map(f(sc.s))(x => (sc withState x._1, x._2) ))
-
-    @inline def retT[M[_], S, A](a: A)(implicit M: Applicative[M]): ReactST[M, S, A] =
-      StateT[M, StateAndCallbacks[S], A](s => M.point((s, a)))
-
-    @inline def retM[M[_], S, A](ma: M[A])(implicit F: Functor[M]): ReactST[M, S, A] = getsT[M, S, A](_ => ma)
-
-    @inline def getT[M[_]: Applicative, S]: ReactST[M, S, S] = get.lift[M]
-
-    def getsT[M[_], S, A](f: S => M[A])(implicit F: Functor[M]): ReactST[M, S, A] =
-      StateT[M, StateAndCallbacks[S], A](sc => F.map(f(sc.s))((sc, _)))
-
-    @inline def setM[M[_]: Functor, S](ms: M[S]): ReactST[M, S, Unit] = modT((_: S) => ms)
-
-    @inline def setT[M[_]: Applicative, S](s: S): ReactST[M, S, Unit] = set(s).lift[M]
-
-    def modT[M[_], S](f: S => M[S])(implicit M: Functor[M]): ReactST[M, S, Unit] =
-      StateT[M, StateAndCallbacks[S], Unit](sc => M.map(f(sc.s))(s2 => (sc withState s2,()) ))
 
     def callbackM[M[_], S, A](ma: M[A], c: OpCallbackIO)(implicit M: Functor[M]): ReactST[M, S, A] =
       StateT[M, StateAndCallbacks[S], A](s => M.map(ma)(a => (s addCallback c, a)))
 
-    @inline def callbackT[M[_]: Applicative, S, A](a: A, c: OpCallbackIO): ReactST[M, S, A] = callback(a, c).lift[M]
+    @inline def gets[S, A](f: S => A): ReactS[S, A] =
+      State.gets[StateAndCallbacks[S], A](s => f(s.s))
 
-    @inline def liftR[M[_]: Monad, S, A](f: S => ReactST[M, S, A]): ReactST[M, S, A] = getT[M, S] flatMap f
+    def getsM[M[_], S, A](f: S => M[A])(implicit F: Functor[M]): ReactST[M, S, A] =
+      StateT[M, StateAndCallbacks[S], A](sc => F.map(f(sc.s))((sc, _)))
 
-    def lift[M[_], S, A](t: StateT[M, S, A])(implicit M: Functor[M]): ReactST[M, S, A] =
+    def liftS[M[_], S, A](t: StateT[M, S, A])(implicit M: Functor[M]): ReactST[M, S, A] =
       StateT[M, StateAndCallbacks[S], A](sc => M.map(t(sc.s))(sa => (sc withState sa._1, sa._2) ))
+
+    def modM[M[_], S](f: S => M[S])(implicit M: Functor[M]): ReactST[M, S, Unit] =
+      StateT[M, StateAndCallbacks[S], Unit](sc => M.map(f(sc.s))(s2 => (sc withState s2,()) ))
+
+    @inline def retT[M[_], S, A](a: A)(implicit M: Applicative[M]): ReactST[M, S, A] =
+      StateT[M, StateAndCallbacks[S], A](s => M.point((s, a)))
 
     def unlift[M[_], S, A](t: ReactST[M, S, A])(implicit M: Functor[M]): StateT[M, S, A] =
       StateT[M, S, A](s => M.map(t(StateAndCallbacks(s)))(sa => (sa._1.s, sa._2) ))
@@ -116,54 +113,61 @@ object ScalazReact {
         M.map(m){ case (sc, a) => (StateAndCallbacks(g(tc.s, sc.s), sc.cb), a) }
       })
 
-    def zoomU[M[_], S, A](r: ReactST[M, Unit, A])(implicit M: Functor[M]): ReactST[M, S, A] =
+    @inline def zoomU[M[_], S, A](r: ReactST[M, Unit, A])(implicit M: Functor[M]): ReactST[M, S, A] =
       zoom[M, Unit, S, A](r, _ => (), (s, _) => s)
 
+    /**
+     * M prefix indicates M in args.
+     * T prefix indicates we want the result lifted into an M.
+     */
     @inline def Fix[S] = new Fix[S]
     final class Fix[S] {
       @inline def apply    [A]     (f: S => (S, A))           : ReactS[S,A]    = ReactS(f)
-      @inline def ret      [A]     (a: A)                     : ReactS[S,A]    = ReactS.ret(a)
+      @inline def callback [A]     (a: A, c: OpCallbackIO)    : ReactS[S,A]    = ReactS.callback(a, c)
       @inline def get                                         : ReactS[S,S]    = ReactS.get
       @inline def gets     [A]     (f: S => A)                : ReactS[S,A]    = ReactS.gets(f)
-      @inline def set              (s: S)                     : ReactS[S,Unit] = ReactS.set(s)
       @inline def mod              (f: S => S)                : ReactS[S,Unit] = ReactS.mod(f)
-      @inline def callback [A]     (a: A, c: OpCallbackIO)    : ReactS[S,A]    = ReactS.callback(a, c)
-      @inline def zoomU    [A]     (r: ReactS[Unit, A])       : ReactS[S,A]    = ReactS.zoomU(r)
+      @inline def ret      [A]     (a: A)                     : ReactS[S,A]    = ReactS.ret(a)
+      @inline def set              (s: S)                     : ReactS[S,Unit] = ReactS.set(s)
       @inline def zoom     [T,A]   (r: ReactS[T, A])
                                    (f: S => T, g: (S, T) => S): ReactS[S,A]    = ReactS.zoom(r, f, g)
+      @inline def zoomU    [A]     (r: ReactS[Unit, A])       : ReactS[S,A]    = ReactS.zoomU(r)
 
-      @inline def applyT   [M[_],A](f: S => M[(S, A)])       (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.applyT(f)
-      @inline def retT     [M[_],A](a: A)                    (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.retT(a)
-      @inline def retM     [M[_],A](ma: M[A])                (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.retM(ma)
-      @inline def getT     [M[_]]                            (implicit M: Applicative[M]): ReactST[M,S,S]    = ReactS.getT
-      @inline def getsT    [M[_],A](f: S => M[A])            (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.getsT(f)
-      @inline def setT     [M[_]]  (s: S)                    (implicit M: Applicative[M]): ReactST[M,S,Unit] = ReactS.setT(s)
-      @inline def setM     [M[_]]  (ms: M[S])                (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.setM(ms)
-      @inline def modT     [M[_]]  (f: S => M[S])            (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.modT(f)
-      @inline def callbackT[M[_],A](a: A, c: OpCallbackIO)   (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.callbackT(a, c)
+      @inline def applyM   [M[_],A](f: S => M[(S, A)])       (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.applyM(f)
       @inline def callbackM[M[_],A](a: M[A], c: OpCallbackIO)(implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.callbackM(a, c)
+      @inline def callbackT[M[_],A](a: A, c: OpCallbackIO)   (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.callbackT(a, c)
+      @inline def getT     [M[_]]                            (implicit M: Applicative[M]): ReactST[M,S,S]    = ReactS.getT
+      @inline def getsM    [M[_],A](f: S => M[A])            (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.getsM(f)
       @inline def liftR    [M[_],A](f: S => ReactST[M, S, A])(implicit M: Monad[M])      : ReactST[M,S,A]    = ReactS.liftR(f)
-      @inline def lift     [M[_],A](t: StateT[M, S, A])      (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.lift(t)
+      @inline def liftS    [M[_],A](t: StateT[M, S, A])      (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.liftS(t)
+      @inline def modM     [M[_]]  (f: S => M[S])            (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.modM(f)
+      @inline def retM     [M[_],A](ma: M[A])                (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.retM(ma)
+      @inline def retT     [M[_],A](a: A)                    (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.retT(a)
+      @inline def setM     [M[_]]  (ms: M[S])                (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.setM(ms)
+      @inline def setT     [M[_]]  (s: S)                    (implicit M: Applicative[M]): ReactST[M,S,Unit] = ReactS.setT(s)
       @inline def unlift   [M[_],A](t: ReactST[M, S, A])     (implicit M: Functor[M])    : StateT [M,S,A]    = ReactS.unlift(t)
     }
 
+    /**
+     * W prefix indicates lack of M in args. (W is M upside-down, woah!)
+     */
     @inline def FixT[M[_], S] = new FixT[M, S]
     final class FixT[M[_], S] {
-      @inline def apply    [A]  (f: S => M[(S, A)])       (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.applyT(f)
-      @inline def applyS   [A]  (f: S => (S, A))          (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS(f).lift[M]
-      @inline def ret      [A]  (a: A)                    (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.retT(a)
-      @inline def retM     [A]  (ma: M[A])                (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.retM(ma)
-      @inline def get                                     (implicit M: Applicative[M]): ReactST[M,S,S]    = ReactS.getT
-      @inline def gets     [A]  (f: S => M[A])            (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.getsT(f)
-      @inline def getsS    [A]  (f: S => A)               (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.gets(f).lift[M]
-      @inline def set           (s: S)                    (implicit M: Applicative[M]): ReactST[M,S,Unit] = ReactS.setT(s)
-      @inline def setM          (ms: M[S])                (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.setM(ms)
-      @inline def mod           (f: S => M[S])            (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.modT(f)
-      @inline def modS          (f: S => S)               (implicit M: Applicative[M]): ReactST[M,S,Unit] = ReactS.mod(f).lift[M]
+      @inline def apply    [A]  (f: S => M[(S, A)])       (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.applyM(f)
+      @inline def applyW   [A]  (f: S => (S, A))          (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.applyT(f)
       @inline def callback [A]  (a: M[A], c: OpCallbackIO)(implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.callbackM(a, c)
-      @inline def callbackS[A]  (a: A, c: OpCallbackIO)   (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.callbackT(a, c)
+      @inline def callbackW[A]  (a: A, c: OpCallbackIO)   (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.callbackT(a, c)
+      @inline def get                                     (implicit M: Applicative[M]): ReactST[M,S,S]    = ReactS.getT
+      @inline def gets     [A]  (f: S => M[A])            (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.getsM(f)
+      @inline def getsW    [A]  (f: S => A)               (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.getsT(f)
       @inline def liftR    [A]  (f: S => ReactST[M, S, A])(implicit M: Monad[M])      : ReactST[M,S,A]    = ReactS.liftR(f)
-      @inline def lift     [A]  (t: StateT[M, S, A])      (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.lift(t)
+      @inline def liftS    [A]  (t: StateT[M, S, A])      (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.liftS(t)
+      @inline def mod           (f: S => M[S])            (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.modM(f)
+      @inline def modW          (f: S => S)               (implicit M: Applicative[M]): ReactST[M,S,Unit] = ReactS.modT(f)
+      @inline def ret      [A]  (ma: M[A])                (implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.retM(ma)
+      @inline def retW     [A]  (a: A)                    (implicit M: Applicative[M]): ReactST[M,S,A]    = ReactS.retT(a)
+      @inline def set           (ms: M[S])                (implicit M: Functor[M])    : ReactST[M,S,Unit] = ReactS.setM(ms)
+      @inline def setW          (s: S)                    (implicit M: Applicative[M]): ReactST[M,S,Unit] = ReactS.setT(s)
       @inline def unlift   [A]  (t: ReactST[M, S, A])     (implicit M: Functor[M])    : StateT [M,S,A]    = ReactS.unlift(t)
       @inline def zoom     [T,A](r: ReactST[M, T, A])
                                 (f: S => T, g: (S,T) => S)(implicit M: Functor[M])    : ReactST[M,S,A]    = ReactS.zoom(r, f, g)
@@ -173,7 +177,7 @@ object ScalazReact {
 
   implicit final class SzRExt_StateTOps[M[_], S, A](val s: StateT[M, S, A]) extends AnyVal {
     @inline def liftR(implicit M: Functor[M]): ReactST[M, S, A] =
-      ReactS lift s
+      ReactS liftS s
   }
 
   implicit final class SzRExt__StateTOps[I, M[_], S, A](val f: I => StateT[M, S, A]) extends AnyVal {