+++ HISTORY.md
@@ -3,8 +3,8 @@ History
 
 ### 0.5.3 (unreleased)
 
-* Workaround for Scala's type inference failing with `StateT.liftR` on functions.
-  Instead of `f(_).liftR`, `f.liftR` is now available and is confirmed to work in `_runState`.
+* Workaround for Scala's type inference failing with `StateT.liftS` on functions.
+  Instead of `f(_).liftS`, `f._liftS` is now available and is confirmed to work in `_runState`.
 
 ### 0.5.2 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.5.1...v0.5.2))